+++ src/main/java/com/potential/hackathon/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.potential.hackathon.controller;
 
-import com.potential.hackathon.model.User;
+import com.potential.hackathon.entity.User;
 import com.potential.hackathon.service.UserService;
 import org.springframework.web.bind.annotation.*;
