+++ src/main/java/com/potential/hackathon/repository/ImageRepository.java
@@ -1,6 +1,6 @@
 package com.potential.hackathon.repository;
 
-import com.potential.hackathon.model.Image;
+import com.potential.hackathon.entity.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageRepository extends JpaRepository<Image, Long> {
