+++ src/main/java/com/potential/hackathon/repository/PostRepository.java
@@ -1,6 +1,6 @@
 package com.potential.hackathon.repository;
 
-import com.potential.hackathon.model.Post;
+import com.potential.hackathon.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
