+++ src/main/java/com/potential/hackathon/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package com.potential.hackathon.repository;
 
 
-import com.potential.hackathon.model.User;
+import com.potential.hackathon.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
